% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_data_retrieval.R
\name{query}
\alias{query}
\title{Performs and SQL query through a provided ODBC connection and return its results as a \code{data.table}}
\usage{
query(connection = DB_IOTDB(), query = "SELECT 1;")
}
\arguments{
\item{connection}{An ODBC connection to a RDBMS server}

\item{query}{The query to perform}
}
\value{
The results of executing \code{query} through \code{connection} as a data table
}
\description{
Performs and SQL query through a provided ODBC connection and return its results as a \code{data.table}
}
\examples{
query(connection = DB_IOTDB(), query = "SELECT * FROM V_LEGACY_NC")
query(connection = DB_IOTCSTATISTICS(), query = "SELECT * FROM CL_FISHING_GROUNDS WHERE CL_FISHING_GROUND_TYPE_ID = 2")
}
