% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_data_retrieval.R
\name{query_IOTDB}
\alias{query_IOTDB}
\title{Performs an SQL query onto an instance of the IOTDB database hosted by a specified MSSQL server and returns its results as a \code{data.table}}
\usage{
query_IOTDB(server = "IOTCS09", query = "SELECT 1;")
}
\arguments{
\item{server}{The server name / IP hosting the IOTDB database}

\item{query}{The query to perform}
}
\value{
A data table with the results of executing \code{query} on the instance of IOTDB hosted by \code{server}
}
\description{
Performs an SQL query onto an instance of the IOTDB database hosted by a specified MSSQL server and returns its results as a \code{data.table}
}
\examples{
query(server = "IOTCS09", query = "SELECT * FROM V_LEGACY_NC")
query(server = "localhost", query = "SELECT * FROM CL_FISHING_GROUNDS WHERE CL_FISHING_GROUND_TYPE_ID = 2")
}
